requirements.txt
# Web framework & server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0

# Banco de Dados
sqlalchemy>=2.0.29

# Migrações de Banco
alembic>=1.13.1

# Segurança e Autenticação
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9

# Testes e qualidade
pytest>=8.1.1
httpx>=0.27.0
ruff>=0.3.5

# Frontend
streamlit>=1.32.0

pydantic-settings>=2.2.1